Use imported_electricity cost to set MO import price
Refactors the ProducerAdapter into a separate DispatchableAdapter class,
and made ImportAdapter inherit from the Producer. Allows the import
interconnector to use the price of the imported electricity carrier in
order to behave like any other dispatchable.

Ref quintel/etmodel#2826
